+++ src/ndarray.cpp
@@ -59,6 +59,11 @@ void NDArray::_bind_methods() {
 	godot::ClassDB::bind_method(D_METHOD("to_packed_byte_array"), &NDArray::to_packed_byte_array);
 	godot::ClassDB::bind_method(D_METHOD("to_packed_int32_array"), &NDArray::to_packed_int32_array);
 	godot::ClassDB::bind_method(D_METHOD("to_packed_int64_array"), &NDArray::to_packed_int64_array);
+	godot::ClassDB::bind_method(D_METHOD("to_packed_vector2_array"), &NDArray::to_packed_vector2_array);
+	godot::ClassDB::bind_method(D_METHOD("to_packed_vector3_array"), &NDArray::to_packed_vector3_array);
+	godot::ClassDB::bind_method(D_METHOD("to_packed_vector4_array"), &NDArray::to_packed_vector4_array);
+	godot::ClassDB::bind_method(D_METHOD("to_packed_color_array"), &NDArray::to_packed_color_array);
+
 	godot::ClassDB::bind_method(D_METHOD("to_godot_array"), &NDArray::to_godot_array);
 
 	godot::ClassDB::bind_method(D_METHOD("assign_add", "a", "b"), &NDArray::assign_add);
@@ -145,7 +150,10 @@ va::DType NDArray::dtype() const {
 }
 
 PackedInt64Array NDArray::shape() const {
-	return packed_from_sequence<PackedInt64Array>(array.shape);
+	PackedInt64Array shape;
+	shape.resize(static_cast<int64_t>(array.shape.size()));
+	std::copy(array.shape.begin(), array.shape.end(), shape.ptrw());
+	return shape;
 }
 
 uint64_t NDArray::size() const {
@@ -291,8 +299,14 @@ PackedFloat32Array NDArray::to_packed_float32_array() const {
 #ifdef NUMDOT_DISABLE_GODOT_CONVERSION_FUNCTIONS
 	throw std::runtime_error("function explicitly disabled; recompile without NUMDOT_DISABLE_GODOT_CONVERSION_FUNCTIONS to enable it.");
 #else
-	ERR_FAIL_COND_V_MSG(array.dimension() > 1, {}, "flatten the array before converting to packed");
-	return varray_to_packed<PackedFloat32Array>(array);
+	ERR_FAIL_COND_V_MSG(array.dimension() != 1, {}, "flatten the array before converting to packed");
+
+	PackedFloat32Array packed;
+	packed.resize(static_cast<int64_t>(array.shape[0]));
+	const auto compute_variant = array.to_compute_variant();
+	fill_c_array_flat(packed.ptrw(), compute_variant);
+
+	return packed;
 #endif
 }
 
@@ -300,8 +314,14 @@ PackedFloat64Array NDArray::to_packed_float64_array() const {
 #ifdef NUMDOT_DISABLE_GODOT_CONVERSION_FUNCTIONS
 	throw std::runtime_error("function explicitly disabled; recompile without NUMDOT_DISABLE_GODOT_CONVERSION_FUNCTIONS to enable it.");
 #else
-	ERR_FAIL_COND_V_MSG(array.dimension() > 1, {}, "flatten the array before converting to packed");
-	return varray_to_packed<PackedFloat64Array>(array);
+	ERR_FAIL_COND_V_MSG(array.dimension() != 1, {}, "flatten the array before converting to packed");
+
+	PackedFloat64Array packed;
+	packed.resize(static_cast<int64_t>(array.shape[0]));
+	const auto compute_variant = array.to_compute_variant();
+	fill_c_array_flat(packed.ptrw(), compute_variant);
+
+	return packed;
 #endif
 }
 
@@ -309,8 +329,14 @@ PackedByteArray NDArray::to_packed_byte_array() const {
 #ifdef NUMDOT_DISABLE_GODOT_CONVERSION_FUNCTIONS
 	throw std::runtime_error("function explicitly disabled; recompile without NUMDOT_DISABLE_GODOT_CONVERSION_FUNCTIONS to enable it.");
 #else
-	ERR_FAIL_COND_V_MSG(array.dimension() > 1, {}, "flatten the array before converting to packed");
-	return varray_to_packed<PackedByteArray>(array);
+	ERR_FAIL_COND_V_MSG(array.dimension() != 1, {}, "flatten the array before converting to packed");
+
+	PackedByteArray packed;
+	packed.resize(static_cast<int64_t>(array.shape[0]));
+	const auto compute_variant = array.to_compute_variant();
+	fill_c_array_flat(packed.ptrw(), compute_variant);
+
+	return packed;
 #endif
 }
 
@@ -318,8 +344,14 @@ PackedInt32Array NDArray::to_packed_int32_array() const {
 #ifdef NUMDOT_DISABLE_GODOT_CONVERSION_FUNCTIONS
 	throw std::runtime_error("function explicitly disabled; recompile without NUMDOT_DISABLE_GODOT_CONVERSION_FUNCTIONS to enable it.");
 #else
-	ERR_FAIL_COND_V_MSG(array.dimension() > 1, {}, "flatten the array before converting to packed");
-	return varray_to_packed<PackedInt32Array>(array);
+	ERR_FAIL_COND_V_MSG(array.dimension() != 1, {}, "flatten the array before converting to packed");
+
+	PackedInt32Array packed;
+	packed.resize(static_cast<int64_t>(array.shape[0]));
+	const auto compute_variant = array.to_compute_variant();
+	fill_c_array_flat(packed.ptrw(), compute_variant);
+
+	return packed;
 #endif
 }
 
@@ -327,8 +359,82 @@ PackedInt64Array NDArray::to_packed_int64_array() const {
 #ifdef NUMDOT_DISABLE_GODOT_CONVERSION_FUNCTIONS
 	throw std::runtime_error("function explicitly disabled; recompile without NUMDOT_DISABLE_GODOT_CONVERSION_FUNCTIONS to enable it.");
 #else
-	ERR_FAIL_COND_V_MSG(array.dimension() > 1, {}, "flatten the array before converting to packed");
-	return varray_to_packed<PackedInt64Array>(array);
+	ERR_FAIL_COND_V_MSG(array.dimension() != 1, {}, "flatten the array before converting to packed");
+
+	PackedInt64Array packed;
+	packed.resize(static_cast<int64_t>(array.shape[0]));
+	const auto compute_variant = array.to_compute_variant();
+	fill_c_array_flat(packed.ptrw(), compute_variant);
+
+	return packed;
+#endif
+}
+
+PackedVector2Array NDArray::to_packed_vector2_array() const {
+#ifdef NUMDOT_DISABLE_GODOT_CONVERSION_FUNCTIONS
+	throw std::runtime_error("function explicitly disabled; recompile without NUMDOT_DISABLE_GODOT_CONVERSION_FUNCTIONS to enable it.");
+#else
+	ERR_FAIL_COND_V_MSG(array.dimension() != 2, {}, "flatten the array before converting to packed");
+	ERR_FAIL_COND_V_MSG(array.shape[1] != 2, {}, "final array dimension must be size 2");
+	// TODO Handle row major/minor? This still assumes it's normal row-major, i think.
+
+	PackedVector2Array packed;
+	packed.resize(static_cast<int64_t>(array.shape[0] * 2));
+	const auto compute_variant = array.to_compute_variant();
+	fill_c_array_flat(&packed.ptrw()->coord[0], compute_variant);
+
+	return packed;
+#endif
+}
+
+PackedVector3Array NDArray::to_packed_vector3_array() const {
+#ifdef NUMDOT_DISABLE_GODOT_CONVERSION_FUNCTIONS
+	throw std::runtime_error("function explicitly disabled; recompile without NUMDOT_DISABLE_GODOT_CONVERSION_FUNCTIONS to enable it.");
+#else
+	ERR_FAIL_COND_V_MSG(array.dimension() != 3, {}, "flatten the array before converting to packed");
+	ERR_FAIL_COND_V_MSG(array.shape[1] != 3, {}, "final array dimension must be size 2");
+	// TODO Handle row major/minor? This still assumes it's normal row-major, i think.
+
+	PackedVector3Array packed;
+	packed.resize(static_cast<int64_t>(array.shape[0] * 3));
+	const auto compute_variant = array.to_compute_variant();
+	fill_c_array_flat(&packed.ptrw()->coord[0], compute_variant);
+
+	return packed;
+#endif
+}
+
+PackedVector4Array NDArray::to_packed_vector4_array() const {
+#ifdef NUMDOT_DISABLE_GODOT_CONVERSION_FUNCTIONS
+	throw std::runtime_error("function explicitly disabled; recompile without NUMDOT_DISABLE_GODOT_CONVERSION_FUNCTIONS to enable it.");
+#else
+	ERR_FAIL_COND_V_MSG(array.dimension() != 3, {}, "flatten the array before converting to packed");
+	ERR_FAIL_COND_V_MSG(array.shape[1] != 3, {}, "final array dimension must be size 2");
+	// TODO Handle row major/minor? This still assumes it's normal row-major, i think.
+
+	PackedVector4Array packed;
+	packed.resize(static_cast<int64_t>(array.shape[0] * 3));
+	const auto compute_variant = array.to_compute_variant();
+	fill_c_array_flat(&packed.ptrw()->components[0], compute_variant);
+
+	return packed;
+#endif
+}
+
+PackedColorArray NDArray::to_packed_color_array() const {
+#ifdef NUMDOT_DISABLE_GODOT_CONVERSION_FUNCTIONS
+	throw std::runtime_error("function explicitly disabled; recompile without NUMDOT_DISABLE_GODOT_CONVERSION_FUNCTIONS to enable it.");
+#else
+	ERR_FAIL_COND_V_MSG(array.dimension() != 3, {}, "flatten the array before converting to packed");
+	ERR_FAIL_COND_V_MSG(array.shape[1] != 3, {}, "final array dimension must be size 2");
+	// TODO Handle row major/minor? This still assumes it's normal row-major, i think.
+
+	PackedColorArray packed;
+	packed.resize(static_cast<int64_t>(array.shape[0] * 3));
+	const auto compute_variant = array.to_compute_variant();
+	fill_c_array_flat(&packed.ptrw()->components[0], compute_variant);
+
+	return packed;
 #endif
 }